+++ nts.uk/uk.at/at.screen/nts.uk.screen.at.app/src/main/java/nts/uk/screen/at/app/ktgwidget/ktg004/RemainingNumberInforDto.java
@@ -42,6 +42,6 @@ public class RemainingNumberInforDto {
 	private GeneralDate grantDate;
 		
 	// 付与日数
-	private int grantDays;
+	private double grantDays;
 	
 }
